Hibernate DAO implementation has been initially made for Advertisements
git-svn-id: https://n-exchange.googlecode.com/svn/trunk@30 37479ede-2111-a82d-6642-dbc92489a3ae

+++ src/main/java/it/sevenbits/dao/AdvertisementDao.java
@@ -1,7 +1,6 @@
 package it.sevenbits.dao;
 
 import it.sevenbits.entity.hibernate.Advertisement;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
@@ -21,7 +20,7 @@ public interface AdvertisementDao {
      * @param id
      * @return
      */
-    Advertisement findById(Integer id);
+    Advertisement findById(Long id);
 
     /**
      * возвращает все объявления
